==> 1.Prepare_weather_station_id_mapping.ipynb <==
#### In this notebook we will import the original DC rejetive data as a csv file and create the weather station id mapping information. ToDO: hook up with Azure blob container to automatically receive the csv file from the Azure blob container

import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
import pyspark
from functools import reduce
from pyspark.sql import Window
from pyspark.sql import functions as F
from pyspark.sql.functions import desc,asc, monotonically_increasing_id, split, size, length

from IPython.display import display, HTML

%run /Users/p0z00cx/jupter/product_weather_prediction/lib/production_utility.ipynb

data = spark.read.csv('/Users/p0z00cx/hactron_product/rejection_data.csv', header="true", inferSchema="true")
display(data.toPandas())

Update the names of columns

col_list=[
  'item_num', 
  'item_desc', 
  'fineline_num',
  'fineline_desc',
  'dept_desc',
  'vendor_name',
  'vendor_num',
  'inspection_date',
  'inspected_cases',
  'inspected_result',
  'rejection_reason',
  'DC_num',
  'order_date',
  'vendor_9d_num',
  'rejected_cases',
  'minus_quality_cases',
  'country',
  'state',
  'region',
  'po_num',
  'po_line_num',
  'dept_num',
  'avg_core_temp',
  'min_ship_temp',
  'max_ship_temp',
  'dc_name',
  'city_name',
  'state_code',
  'vm_yr_wk',
  'mabd_date',
  'not_ship_date',
  'process_days',
  'pack_day',
  'manual_ind',
  'import_ind',
  'pymnt_typ',
  'ordered_cases',
  'received_cases',
  'ship_city',
  'ship_state',
  'ship_zip'
 ]

oldColumns = data.schema.names
newColumns = col_list
data = reduce(lambda data, idx: data.withColumnRenamed(oldColumns[idx], newColumns[idx]), range(len(oldColumns)),data)
display(data.toPandas())

data_region  = data.select("state", "region").distinct()
display(data_region .toPandas())

# map the state and region with weather station id
state_station_id = spark.read.csv('/Users/p0z00cx/hactron_product/region_state_stationid.csv', header="true", inferSchema="true",)
display(state_station_id.toPandas())

col_list=['state', 'region', 'station_id']
oldColumns = state_station_id.schema.names
newColumns = col_list
state_station_id = reduce(lambda state_station_id, idx: state_station_id.withColumnRenamed(oldColumns[idx], newColumns[idx]), range(len(oldColumns)), state_station_id)
display(state_station_id.toPandas())

from pyspark.sql.functions import concat, lit
state_station_id = state_station_id.withColumn("state_region_key", concat(state_station_id.state, state_station_id.region))
display(state_station_id.toPandas())

need_data = data.select('item_desc', 'dept_desc', 'vendor_name','inspected_result', 'order_date','country','state','region','dc_name','mabd_date', 'ordered_cases').filter(data['inspected_result'] !='null')
need_data = need_data.withColumn("state_region_key", concat(need_data.state, need_data.region))
need_data = need_data.drop("state").drop("region")
display(need_data.toPandas())

need_data_with_station_id = need_data.join(state_station_id, 'state_region_key', 'left')
need_data_with_station_id = need_data_with_station_id.drop('state_region_key')
display(need_data_with_station_id.toPandas())

need_data_with_station_id = need_data_with_station_id.filter(need_data_with_station_id["mabd_date"].isNotNull())
display(need_data_with_station_id.toPandas())

need_data_with_station_id = need_data_with_station_id.withColumn("start_day", get_start_day_udf("mabd_date"))
display(need_data_with_station_id.toPandas())

need_data_with_station_id = need_data_with_station_id.withColumn("end_day", get_end_day_udf("mabd_date"))
display(need_data_with_station_id.toPandas())

need_data_with_station_id.toPandas().to_csv('/Users/p0z00cx/hactron_product/rejection_data_with_station_id.csv', header="true", index=None)
