==> 2.Join_weather_data.ipynb <==
#### This notebook is join the rejection production data with the weather data 

import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
import pyspark
from IPython.display import display, HTML
from functools import reduce
from pyspark.sql import Window
from pyspark.sql import functions as F
from pyspark.sql.functions import concat, lit
from pyspark.sql.functions import desc,asc, monotonically_increasing_id, split, size, length

from pyspark.sql.types import StringType

%run /Users/p0z00cx/jupter/product_weather_prediction/lib/Feature_engineering_helpers.ipynb

%run /Users/p0z00cx/jupter/product_weather_prediction/lib/classifer_helpers.ipynb

%run /Users/p0z00cx/jupter/product_weather_prediction/lib/production_utility.ipynb

product_pd = spark.read.csv('/Users/p0z00cx/hactron_product/rejection_data_with_station_id.csv', header="true", inferSchema="true")
display(product_pd.toPandas())

weather_data_2020 = spark.read.csv('/Users/p0z00cx/Downloads/2020.csv', header="true", inferSchema="true")
weather_data_2019= spark.read.csv('/Users/p0z00cx/Downloads/2019.csv.gz', header="true", inferSchema="true")
weather_data = weather_data_2019.unionAll(weather_data_2020)
weather_data.count()

col_list=['stationId', 'date', 'type', 'value', 'item1', 'item2','item3','item4']
oldColumns = weather_data.schema.names
newColumns = col_list
weather_data = reduce(lambda weather_data, idx: weather_data.withColumnRenamed(oldColumns[idx], newColumns[idx]), range(len(oldColumns)), weather_data)
weather_data.show(10)

#### Get the PRCP

join_condition = [product_pd.station_id == weather_data.stationId,
product_pd.start_day < weather_data.date,
product_pd.end_day > weather_data.date,
weather_data.type == "PRCP"
]

product_pd = product_pd.withColumn("id", monotonically_increasing_id())
product_pd.count()

df3 = product_pd.join(weather_data, how='left', on=join_condition)
df3 = df3.filter("stationId is not null")
df3.count()

df4 = df3.groupBy(['id']).agg(F.collect_list('value').alias('prcp_list'))
df4 = df4.withColumnRenamed("id", "id4")
display(df4)

df4 = df4.withColumn("prcp_aver", get_average_udf("prcp_list"))

df4 = df4.drop("prcp_list")
df4.show(10)

product_pd_final = product_pd.join(df4, how="left", on=[product_pd.id == df4.id4])
product_pd_final = product_pd_final.drop("prcp_list").drop("id4")
product_pd_final.show(5)
product_pd_final.count()

#### get TAVG_list



join_condition = [product_pd.station_id == weather_data.stationId,
product_pd.start_day < weather_data.date,
product_pd.end_day > weather_data.date,
#weather_data.type == "TMIN" ,
#weather_data.type == "TMAX",
weather_data.type == "TAVG"
]

df3 = product_pd.join(weather_data, how='left', on=join_condition)
df3 = df3.filter("stationId is not null")
df3.count()
df3.show(3)

df4 = df3.groupBy(['id']).agg(F.collect_list('value').alias('tavg_list'))
df4 = df4.withColumnRenamed("id", "id4")
df4.count()

df4 = df4.withColumn("tavg", get_average_udf("tavg_list"))
product_pd_final = product_pd_final.join(df4, how="left", on=[product_pd_final.id == df4.id4])
product_pd_final = product_pd_final.drop("tavg_list").drop("id4")
product_pd_final.show(5)
product_pd_final.count()

product_pd_final = product_pd_final.fillna({'prcp_aver': '0'}).fillna({'tavg': '70'})
display(product_pd_final.toPandas())

test_data_set =  product_pd_final.select("item_desc", "dept_desc", "state","region","inspected_result","dc_name","prcp_aver", "tavg")
test_data_set = test_data_set.withColumn("result", udf(lambda x : 0 if x =="rejected" else 1)("inspected_result"))
display(test_data_set.toPandas())

test_data_set.toPandas().to_csv('/Users/p0z00cx/hactron_product/test_data_set.csv', header="true", index=None)
