==> 3b.random_forest_tree_training.ipynb <==
# DictVectorizer
from sklearn.feature_extraction import DictVectorizer

# XGB
import xgboost as xgb

# cross_val_score
from sklearn.model_selection import cross_val_score

# Import pandas
import pandas as pd


from sklearn.compose import make_column_transformer
from sklearn.preprocessing import OneHotEncoder
from sklearn.linear_model import LogisticRegression
from sklearn.pipeline import make_pipeline
from sklearn.model_selection import cross_val_score

from pyspark.sql.functions import desc,asc, monotonically_increasing_id, split, size, length

# Plot AUC *(Self Defined Package)
from sklearn.metrics import roc_curve, auc, roc_auc_score

%run /Users/p0z00cx/jupter/product_weather_prediction/lib/production_utility.ipynb

%run /Users/p0z00cx/jupter/product_weather_prediction/lib/roc_auc.ipynb

test_data_set = spark.read.csv('/Users/p0z00cx/hactron_product/test_data_set_final.csv', header="true", inferSchema="true")
display(test_data_set.toPandas())

df = test_data_set.toPandas()
df.isna().sum()

X= df.loc[:,['dept_desc_new','state_new','dc_name','prcp_aver','tavg']]
X = pd.get_dummies(X, prefix_sep='_', drop_first=True)
X

Y = df.result  
Y

# Instantiate XGBClassifier
from sklearn.metrics import classification_report
xgb_cl = xgb.XGBClassifier()
clf_1 = xgb_cl.fit(X,Y)
pred_y_1 = clf_1.predict(X)
target_names = ['result 0', 'result 1']
print (classification_report( Y,pred_y_1, target_names=target_names))
